restructured to cookiecutter template of pelican #5
