+++ .gitignore
@@ -54,3 +54,5 @@ docs/_build/
 
 # project specific
 .idea/
+
+poetry.lock
