+++ CHANGELOG.md
@@ -16,4 +16,3 @@
 -----
 
 * initial release
-
