+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+Contributing
+============
+
+Contributions are welcome and much appreciated. Every little bit helps. You can contribute by improving the documentation, adding missing features, and fixing bugs. You can also help out by reviewing and commenting on [existing issues][].
+
+To start contributing to this plugin, review the [Contributing to Pelican][] documentation, beginning with the **Contributing Code** section.
+
+[existing issues]: https://github.com/pelican-plugins/pelican-minification/issues
+[Contributing to Pelican]: https://docs.getpelican.com/en/latest/contribute.html
