+++ DEVELOPMENT.md
@@ -1,0 +1,13 @@
+# Development
+
+## Prerequisites
+Install prerequisites:
+
+```
+python3 -m pip install poetry pre-commit
+pre-commit install
+poetry install
+```
+## Commands
+* `poetry run invoke test`
+* `poetry run invoke lint --diff`
