+++ FUNDING.yml
@@ -1,0 +1,2 @@
+custom: https://donate.getpelican.com
+liberapay: pelican
