--- setup.cfg
@@ -1,49 +1,0 @@
-[metadata]
-name = pelican-minification
-version = attr: pelican.plugins.minification.__version__
-description = Minifies HTML, CSS and JS of generated Pelican content.
-long_description = file: README.rst
-long_description_content_type = text/x-rst
-author = Alexander Herrmann
-author_email = github@torv.rocks
-url = https://github.com/dArignac/pelican-minification
-
-license_files = LICENSE
-keywords = pelican, javascript, css, html, minification
-classifiers =
-    Development Status :: 5 - Production/Stable
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    Intended Audience :: End Users/Desktop
-    Natural Language :: English
-    Operating System :: POSIX :: Linux
-    Topic :: Text Processing :: Markup
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    License :: OSI Approved :: MIT License
-
-[options]
-zip_safe = False
-include_package_data = True
-packages = pelican.plugins.minification
-install_requires =
-    beautifulsoup4
-    csscompressor>=0.9.5
-    htmlmin>=0.1.12
-    jsmin>=3.0.1
-    packaging
-    pelican>=4
-
-[options.extras_require]
-dev =
-    prospector[with_pyroma]
-    pytest-cov>=2.6.1
-    pytest>=5.2.0
-    pytest-runner
-
-[aliases]
-test=pytest
