--- setup.py
@@ -1,2 +1,0 @@
-from setuptools import setup
-setup()
