+++ tox.ini
@@ -1,0 +1,3 @@
+[flake8]
+max-line-length = 88
+ignore = E203, W503
